Refactor: ui 로직과 form 로직 분리 [#219]
