+++ src/hooks/index.ts
@@ -1,3 +1,4 @@
 export { default as useComponentContext } from './useComponentContext';
 export { default as useToastOpen } from './useToastOpen';
 export { default as useToolBanner } from './useToolBanner';
+export { default as useReport } from './useReport';
